requirements.txt
scikit-learn
pandas
numpy
scipy
xgboost
sqlalchemy
psycopg2-binary
boto3
fastapi
uvicorn
python-multipart
jinja2
evidently
python-dotenv
pyyaml
dill
-e .
matplotlib
seaborn
mypy-boto3-s3